Code refactoring and started statistics
